+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-18 11:51:48
-// Commit: fix(ui): fix memory leak by cancelling ui subscription
+// Updated at: 2026-04-18 18:38:46
+// Commit: docs(ui): update readme and API contracts for ui
 
 class uiComponentState {
-  final String commitRef = '509';
-  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-18T11:51:48+05:00');
+  final String commitRef = '528';
+  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
+  final DateTime timestamp = DateTime.parse('2026-04-18T18:38:46+05:00');
 }